custom_components/smart_irrigation/requirements.test.txt
pytest
pytest-cov
pytest-homeassistant-custom-component